Implement review findings
